switch from a git archive to borg backup -- better compression but totally stupid
This may be a huge paint point: git and git fat are not great, and
git-fat in particular doesn't do deduplication of similar or even identical
files.

Borg Backup does -- but it's super wonky and not well-supported on windows.
We'll have to see how it does. Might switch back and also open to LOTS
of other alternatives, if they exist.

git/git-fat have advantage that it's a one-liner to get a .zip file.
borg-backup it's easy to restore the *folder*, and deduplication is rad.